mm,thp,shmem: make tmpfs mount flags enough to enable khugepaged
Currently if thp enabled=[madvise], mounting a tmpfs filesystem
with huge=always and mmapping files from that tmpfs does not
result in khugepaged collapsing those mappings, despite the
mount flag indicating that it should.

Fix that by breaking up the blocks of tests in hugepage_vma_check
a little bit, and testing things in order.

Signed-off-by: Rik van Riel <riel@surriel.com>