btrfs: warn when remount will not change the free space tree
Summary:
btrfs: warn when remount will not change the free space tree

If the remount is ro->ro, rw->ro, or rw->rw, we will not create or
clear the free space tree. This can be surprising, so print a warning
to dmesg to make the failure more visible. It is also important to
ensure that the space cache options (SPACE_CACHE, FREE_SPACE_TREE) are
consistent, so ensure those are set to properly match the current on
disk state (which won't be changing).


(cherry picked from commit 2838d255cb9b85a845efc3bbd3f6fc66ed883d35)

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Boris Burkov <boris@bur.io>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>