tcp: Set ECT0 bit in tos/tclass for synack when BPF needs ECN
Summary:
tcp: Set ECT0 bit in tos/tclass for synack when BPF needs ECN

When a BPF program is used to select between a type of TCP congestion
control algorithm that uses either ECN or not there is a case where the
synack for the frame was coming up without the ECT0 bit set. A bit of
research found that this was due to the final socket being configured to
dctcp while the listener socket was staying in cubic.

To reproduce it all that is needed is to monitor TCP traffic while running
the sample bpf program "samples/bpf/tcp_cong_kern.c". What is observed,
assuming tcp_dctcp module is loaded or compiled in and the traffic matches
the rules in the sample file, is that for all frames with the exception of
the synack the ECT0 bit is set.

To address that it is necessary to make one additional call to
tcp_bpf_ca_needs_ecn using the request socket and then use the output of
that to set the ECT0 bit for the tos/tclass of the packet.

Fixes: 91b5b21c7c16 ("bpf: Add support for changing congestion control")

Link: https://lore.kernel.org/r/160593039663.2604.1374502006916871573.stgit@localhost.localdomain
Signed-off-by: Jakub Kicinski <kuba@kernel.org>

As a part of the backporting process for this I had to pull in bits from
two other patches:
861602b57730 ("tcp: Allow full IP tos/IPv6 tclass to be reflected in L3 header")`
ac8f1710c12b ("tcp: reflect tos value received in SYN to the socket")`

(cherry picked from commit 407c85c7ddd6b84d3cbdd2275616f70c27c17913)

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Alexander Duyck <alexanderduyck@fb.com>