bnxt_en: Rearrange the logic in bnxt_flash_package_from_fw_obj().
This function will be modified in the next patch to retry flashing
the firmware in a loop.  To facilate that, we rearrange the code so
that the steps that only need to be done once before the loop will be
moved to the top of the function.

Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit a9094ba6072bfabe93f93b641a3858d9c91c2c86)
Signed-off-by: Jonathan Lemon <jonathan.lemon@gmail.com>